Better playback event handling
Playback events used to have a more convoluted process of being emitted back to everyone, including the user who sent out the playback event, which required more complex handling of playback updates. This is no longer the case as events are not emitted back to the user who sent the playback update.